split npu_instance
