+++ paddle/fluid/platform/CMakeLists.txt
@@ -140,6 +140,10 @@ if(WITH_ASCEND_CL)
     target_link_libraries(device_context npu_resource_pool)
 endif()
 
+IF(WITH_ASCEND_CL)
+cc_library(npu_instance SRCS npu_instance.cc DEPS npu_info device_context) 
+ENDIF()
+
 nv_test(device_context_test SRCS device_context_test.cu DEPS device_context gpu_info)
 
 cc_test(init_test SRCS init_test.cc DEPS device_context)
