+++ paddle/fluid/platform/npu_info.cc
@@ -371,47 +371,5 @@ bool IsNPUMallocRecorded(int dev_id) {
   return RecordedNPUMallocHelper::Instance(dev_id)->NeedRecord();
 }
 
-AclInstance::~AclInstance() {}
-
-AclInstance &AclInstance::Instance() {
-  static AclInstance instance;
-  return instance;
-}
-
-AclInstance::AclInstance() {
-  PADDLE_ENFORCE_NPU_SUCCESS(aclInit(nullptr));
-  VLOG(4) << "Call aclrtSetDevice ";
-  // NOTE(zhiqiu): why set devices here?
-  // Because ACL creates a default context which contains 2 streams
-  // when calling aclrtSetDeviceId, so usually we do not need to
-  // create contexts explicitly. And, for each device, aclrtSetDeviceId
-  // need to call parily with aclrtResetDeviceId to destory the default
-  // context. Here, we use this singleton and static instance to manage
-  // the devices to make sure they will be resetted before program exit.
-  devices_ = platform::GetSelectedNPUDevices();
-  for (auto it = devices_.rbegin(); it != devices_.rend(); ++it) {
-    SetNPUDeviceId(*it);
-    VLOG(4) << "Call aclrtSetDevice " << *it;
-  }
-}
-
-void AclInstance::Finalize() {
-  // NOTE(zhiqiu): DO NOT perform finalize in destructor
-  // to avoid problems caused by destructor order of static
-  // object.
-  for (size_t i = 0; i < devices_.size(); ++i) {
-    auto status = aclrtResetDevice(devices_[i]);
-    VLOG(4) << "Call aclrtResetDevice " << devices_[i]
-            << " status = " << status;
-    auto dev_ctx = platform::DeviceContextPool::Instance().Get(
-      platform::NPUPlace(devices_[i])));
-    auto npu_stream =
-        static_cast<platform::NPUDeviceContext *>(dev_ctx)->NPUstream();
-    npu_stream->Destroy();
-  }
-  auto status = aclFinalize();
-  VLOG(4) << "Call aclFinalize, status = " << status;
-}
-
 }  // namespace platform
 }  // namespace paddle
